fix: 편집 정산이 gesture 커밋을 양보 뒤로 미뤄 선택 변경에 취소되던 문제
